extracted another toy network
