+++ network_diffusion/networks/toy_network_cim.py
@@ -1,3 +1,11 @@
+# Copyright (c) 2025 by Michał Czuba, Piotr Bródka.
+#
+# This file is a part of Network Diffusion.
+#
+# Network Diffusion is licensed under the MIT License. You may obtain a copy
+# of the License at https://opensource.org/licenses/MIT
+# =============================================================================
+
 """A toy network used with by the authors of CIM method."""
 
 import networkx as nx
